LFLT-361 Sitemap should be exposed to LCFA via Bridge  - Aligned sitemap controller so it can serialize sitemap as JSON besides XML  - Removed Sitemap model class (moved to REST API package)  - Updated REST API version to v1.6.4  - Updated RCP version to v1.5.3  - Aligned tests
