+++ acceptance-tests/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>leaflet</artifactId>
         <groupId>hu.psprog.leaflet</groupId>
-        <version>1.9.3-dev</version>
+        <version>1.9.4-dev</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
