+++ persistence/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>hu.psprog.leaflet</groupId>
         <artifactId>leaflet</artifactId>
-        <version>1.9.3-dev</version>
+        <version>1.9.4-dev</version>
     </parent>
     
     <artifactId>leaflet-persistence</artifactId>
