+++ pom.xml
@@ -7,7 +7,7 @@
     <groupId>hu.psprog.leaflet</groupId>
     <artifactId>leaflet</artifactId>
     <packaging>pom</packaging>
-    <version>1.9.3-dev</version>
+    <version>1.9.4-dev</version>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -40,10 +40,10 @@
         <java.version>11</java.version>
 
         <!-- leaflet internal dependency versions -->
-        <leaflet.rest-api.version>1.6.3</leaflet.rest-api.version>
+        <leaflet.rest-api.version>1.6.4</leaflet.rest-api.version>
         <leaflet.jwt-component.version>2.4.0</leaflet.jwt-component.version>
         <leaflet.mailer-component.version>1.2.1</leaflet.mailer-component.version>
-        <leaflet.rcp.version>1.5.2</leaflet.rcp.version>
+        <leaflet.rcp.version>1.5.3</leaflet.rcp.version>
         <leaflet.bridge.version>3.2.2</leaflet.bridge.version>
         <leaflet.tlp-appender.version>1.1.0</leaflet.tlp-appender.version>
         <leaflet.translation-adapter.version>2.1.0</leaflet.translation-adapter.version>
